build: make Stop-DevPowerShell only Resolve-Path once

+++ build.psm1
@@ -254,11 +254,15 @@ function Start-PSBuild {
         throw "Cross compiling for win-arm or win-arm64 is only supported on Windows environment"
     }
     function Stop-DevPowerShell {
+        if (-not (Test-Path $script:Options.Output)) {
+            return
+        }
+        $pwshDev = Resolve-Path $script:Options.Output
         Get-Process pwsh* |
             Where-Object {
                 $_.Modules |
                 Where-Object {
-                    $_.FileName -eq (Resolve-Path $script:Options.Output).Path
+                    $_.FileName -eq $pwshDev.Path
                 }
             } |
         Stop-Process -Verbose
